Wait for SSH command to complete after execution
This patch resolves a problem that started to show up when commands,
such as 'cat' were executed on target. The original code was killing the
command once we faced a return_value == None. This is obviously wrong
and it probably comes from multiple SSH module iterations.

The right thing to do is to wait for the completion of the process, once
we terminate reading its stdout. In this way, return_value will be !=
None and we can process fast commands execution.