session: gracefully stop on interrupts
Remove tasks cancel because it's simply not working for our scenario. By
cancelling all the tasks, we just get rid of the current running
coroutines, but the overall application is stopping without killing, for
example, tests executables.
Now we just wait for the right exceptions and stop the session in a
graceful way.

Signed-off-by: Andrea Cervesato <andrea.cervesato@suse.com>